chore: split EventManager operations into EventProcessor and EventUploader
